+++ src/main/java/com/practise/project/repository/DepartmentRepository.java
@@ -12,7 +12,7 @@ public interface DepartmentRepository extends JpaRepository<Department,Integer> 
 
 	Optional<Department> findByDeptCodeAndActive(String deptCode, boolean active);
 
-	Optional<Department> findByIdAndActive(Integer id, boolean b);
+	Optional<Department> findByIdAndActive(Long id, boolean b);
 
 	Page<Department> findByActive(boolean b, Pageable pageableInstance);
